ref(*): move awsclient to pkg
