+++ pkg/aws/route53_test.go
@@ -1,4 +1,4 @@
-package awsclient
+package aws
 
 import (
 	"testing"
